+++ mayan/apps/dynamic_search/tests/test_models.py
@@ -3,33 +3,23 @@ from __future__ import unicode_literals
 from django.test import override_settings
 
 from common.tests import BaseTestCase
-from documents.models import DocumentType
 from documents.search import document_search
-from documents.tests import TEST_DOCUMENT_TYPE_LABEL, TEST_SMALL_DOCUMENT_PATH
+from documents.tests import (
+    DocumentTestMixin, TEST_DOCUMENT_FILENAME, TEST_SMALL_DOCUMENT_FILENAME
+)
 
 
 @override_settings(OCR_AUTO_OCR=False)
-class DocumentSearchTestCase(BaseTestCase):
-    def setUp(self):
-        super(DocumentSearchTestCase, self).setUp()
-        self.document_type = DocumentType.objects.create(
-            label=TEST_DOCUMENT_TYPE_LABEL
-        )
-
-        with open(TEST_SMALL_DOCUMENT_PATH) as file_object:
-            self.document = self.document_type.new_document(
-                file_object=file_object, label='mayan_11_1.pdf'
-            )
-
-    def tearDown(self):
-        self.document_type.delete()
-        super(DocumentSearchTestCase, self).tearDown()
+class DocumentSearchTestCase(DocumentTestMixin, BaseTestCase):
+    auto_upload_document = False
+    test_document_filename = TEST_DOCUMENT_FILENAME
 
     def test_simple_search_after_related_name_change(self):
         """
         Test that simple search works after related_name changes to
         document versions and document version pages
         """
+        self.document = self.upload_document()
         queryset, elapsed_time = document_search.search(
             {'q': 'Mayan'}, user=self.admin_user
         )
@@ -38,6 +28,7 @@ class DocumentSearchTestCase(BaseTestCase):
 
     def test_advanced_search_after_related_name_change(self):
         # Test versions__filename
+        self.document = self.upload_document()
         queryset, elapsed_time = document_search.search(
             {'label': self.document.label}, user=self.admin_user
         )
@@ -53,10 +44,11 @@ class DocumentSearchTestCase(BaseTestCase):
         self.assertTrue(self.document in queryset)
 
     def test_simple_or_search(self):
-        with open(TEST_SMALL_DOCUMENT_PATH) as file_object:
-            self.document_2 = self.document_type.new_document(
-                file_object=file_object, label='second_doc.pdf'
-            )
+        self.document = self.upload_document()
+        self.test_document_filename = TEST_SMALL_DOCUMENT_FILENAME
+        self.document_2 = self.upload_document()
+        self.document_2.label = 'second_doc.pdf'
+        self.document_2.save()
 
         queryset, elapsed_time = document_search.search(
             {'q': 'Mayan OR second'}, user=self.admin_user
@@ -66,12 +58,48 @@ class DocumentSearchTestCase(BaseTestCase):
         self.assertTrue(self.document_2 in queryset)
 
     def test_simple_and_search(self):
-        with open(TEST_SMALL_DOCUMENT_PATH) as file_object:
-            self.document_2 = self.document_type.new_document(
-                file_object=file_object, label='second_doc.pdf'
-            )
+        self.test_document_filename = TEST_SMALL_DOCUMENT_FILENAME
+        self.document_2 = self.upload_document()
+        self.document_2.label = 'second_doc.pdf'
+        self.document_2.save()
 
         queryset, elapsed_time = document_search.search(
-            {'q': 'Mayan second'}, user=self.admin_user
+            {'q': 'non_valid second'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 0)
+
+        queryset, elapsed_time = document_search.search(
+            {'q': 'second non_valid'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 0)
+
+    def test_simple_negated_search(self):
+        self.test_document_filename = TEST_SMALL_DOCUMENT_FILENAME
+        self.document_2 = self.upload_document()
+        self.document_2.label = 'second_doc.pdf'
+        self.document_2.save()
+
+        queryset, elapsed_time = document_search.search(
+            {'q': '-non_valid second'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 1)
+
+        queryset, elapsed_time = document_search.search(
+            {'label': '-second'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 0)
+
+        queryset, elapsed_time = document_search.search(
+            {'label': '-second -Mayan'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 0)
+
+        queryset, elapsed_time = document_search.search(
+            {'label': '-second OR -Mayan'}, user=self.admin_user
+        )
+        self.assertEqual(queryset.count(), 1)
+
+        queryset, elapsed_time = document_search.search(
+            {'label': '-non_valid -second'}, user=self.admin_user
         )
         self.assertEqual(queryset.count(), 0)